refactor(code_gen): extract methods/module.rs
